test(br-utils): update files namespaces
