+++ packages/br-utils/tests/phpunit/Cpf/CpfGeneratorClassTest.php
@@ -2,22 +2,28 @@
 
 declare(strict_types=1);
 
-namespace Lacus\BrUtils\Tests\Cpf;
+namespace Lacus\BrUtils\Tests\Legacy\Cpf;
 
-use function Lacus\BrUtils\Cpf\cpf_gen;
-
+use Lacus\BrUtils\Cpf\CpfGenerator;
 use Lacus\CpfGen\Tests\CpfGeneratorTestCases;
 use PHPUnit\Framework\TestCase;
 
-class CpfGeneratorFunctionTest extends TestCase
+class CpfGeneratorClassTest extends TestCase
 {
     use CpfGeneratorTestCases;
+
+    private CpfGenerator $generator;
+
+    protected function setUp(): void
+    {
+        $this->generator = new CpfGenerator();
+    }
 
     protected function generate(
         ?bool $format = null,
         ?string $prefix = null,
     ): string {
-        return cpf_gen(
+        return $this->generator->generate(
             $format,
             $prefix,
         );
