+++ packages/br-utils/tests/phpunit/Cpf/CpfValidatorFunctionTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Lacus\BrUtils\Tests\Cpf;
+namespace Lacus\BrUtils\Tests\Legacy\Cpf;
 
 use function Lacus\BrUtils\Cpf\cpf_val;
